Made the frames used by AddClass reuseable
