+++ app/src/main/java/org/View/GUICmp/AddClassInput.java
@@ -1,16 +1,8 @@
 package org.View.GUICmp;
 
-import java.awt.Component;
-import java.awt.FlowLayout;
-import java.awt.GridLayout;
-import java.util.ArrayList;
-import java.util.HashMap;
-
-import javax.swing.BorderFactory;
 import javax.swing.BoxLayout;
 import javax.swing.JButton;
 import javax.swing.JFrame;
-import javax.swing.JLabel;
 import javax.swing.JPanel;
 
 import org.Controller.UMLEditor;
@@ -62,207 +54,4 @@ public class AddClassInput extends JFrame {
 		pack();
 	}
 
-	private class NameFrame extends JPanel {
-		private HintTextArea txt = new HintTextArea("Name");
-
-		public NameFrame() {
-			super();
-			setBorder(BorderFactory.createTitledBorder("Name"));
-			add(txt);
-		}
-
-		public String getData() {
-			return txt.getText();
-		}
-
-		public void reset() {
-			txt.reset();
-		}
-	}
-
-	private class FieldFrame extends JPanel {
-		private JPanel grid;
-
-		public FieldFrame() {
-			super();
-			setBorder(BorderFactory.createTitledBorder("Fields"));
-			setLayout(new BoxLayout(this, BoxLayout.Y_AXIS));
-
-			grid = new JPanel();
-			grid.setLayout(new GridLayout(0, 2, 5, 5));
-			add(grid);
-
-			JPanel control = new JPanel();
-			control.setLayout(new BoxLayout(control, BoxLayout.X_AXIS));
-
-			JButton tmp;
-			tmp = new JButton("+");
-			tmp.addActionListener(e -> {
-				GridLayout g = (GridLayout) grid.getLayout();
-				g.setRows(g.getRows() + 1);
-				grid.add(new HintTextArea("Type"));
-				grid.add(new HintTextArea("Name"));
-				grid.revalidate();
-				grid.repaint();
-				pack();
-			});
-			control.add(tmp);
-
-			tmp = new JButton("-");
-			tmp.addActionListener(e -> {
-				GridLayout g = (GridLayout) grid.getLayout();
-				if (g.getRows() < 1)
-					return;
-				g.setRows(g.getRows() - 1);
-				int cnt = grid.getComponentCount();
-				grid.remove(grid.getComponent(cnt - 1));
-				grid.remove(grid.getComponent(cnt - 2));
-				grid.revalidate();
-				grid.repaint();
-				pack();
-			});
-			control.add(tmp);
-			add(control);
-		}
-
-		public HashMap<String, String> getData() {
-			HashMap<String, String> h = new HashMap<>();
-			int cnt = grid.getComponentCount();
-			if (cnt < 2)
-				return h;
-			for (int i = 0; i < cnt; i += 2) {
-				String type, name;
-				HintTextArea a = (HintTextArea) grid.getComponent(i);
-				type = a.getText();
-				a = (HintTextArea) grid.getComponent(i + 1);
-				name = a.getText();
-				h.put(name, type);
-			}
-			return h;
-		}
-
-		public void reset() {
-			GridLayout l = (GridLayout) grid.getLayout();
-			if (l.getRows() < 1)
-				return;
-			grid.removeAll();
-			l.setRows(0);
-		}
-	}
-
-	private class MethodFrame extends JPanel {
-		private JPanel mar;
-
-		public MethodFrame() {
-			super();
-			setBorder(BorderFactory.createTitledBorder("Methods"));
-			setLayout(new BoxLayout(this, BoxLayout.Y_AXIS));
-			JPanel mar = new JPanel();
-			mar.setLayout(new BoxLayout(mar, BoxLayout.Y_AXIS));
-			add(mar);
-			this.mar = mar;
-
-			JPanel control = new JPanel();
-
-			JButton tmp = new JButton("+");
-			tmp.addActionListener(e -> {
-				addMethod(mar);
-				revalidate();
-				repaint();
-				pack();
-			});
-			control.add(tmp);
-			tmp = new JButton("-");
-			tmp.addActionListener(e -> {
-				int cnt = mar.getComponentCount();
-				if (cnt < 1)
-					return;
-				mar.remove(mar.getComponent(cnt - 1));
-				revalidate();
-				repaint();
-				pack();
-			});
-			control.add(tmp);
-			add(control);
-		}
-
-		private void addMethod(JPanel mar) {
-			JPanel method = new JPanel();
-			method.setLayout(new FlowLayout(FlowLayout.LEFT, 5, 5));
-
-			method.add(new ParamFrame());
-
-			JButton tmp = new JButton("+");
-			tmp.addActionListener(e -> {
-				ParamFrame f = (ParamFrame) method.getComponent(0);
-				f.addBox();
-			});
-			method.add(tmp);
-
-			tmp = new JButton("-");
-			tmp.addActionListener(e -> {
-				ParamFrame f = (ParamFrame) method.getComponent(0);
-				f.remBox();
-			});
-			method.add(tmp);
-			mar.add(method);
-		}
-
-		public ArrayList<ArrayList<String>> getData() {
-			ArrayList<ArrayList<String>> r = new ArrayList<>();
-			for (Component c : mar.getComponents()) {
-				JPanel tmp = (JPanel) c;
-				ParamFrame p = (ParamFrame) tmp.getComponent(0);
-				r.add(p.getData());
-			}
-			return r;
-		}
-
-		public void reset() {
-			mar.removeAll();
-		}
-	}
-
-	private class ParamFrame extends JPanel {
-		public ParamFrame() {
-			super();
-			setLayout(new FlowLayout(FlowLayout.RIGHT, 5, 5));
-			add(new HintTextArea("Return type"));
-			add(new HintTextArea("Name"));
-			add(new JLabel("("));
-			add(new JLabel(")"));
-
-		}
-
-		public void addBox() {
-			remove(getComponent(getComponentCount() - 1));
-			add(new HintTextArea("Type"));
-			add(new HintTextArea("Name"));
-			add(new JLabel(")"));
-			pack();
-		}
-
-		public void remBox() {
-			int cnt = getComponentCount();
-			if (cnt <= 4)
-				return;
-			Component a, b;
-			a = getComponent(getComponentCount() - 2);
-			b = getComponent(getComponentCount() - 3);
-			remove(a);
-			remove(b);
-			pack();
-		}
-
-		public ArrayList<String> getData() {
-			ArrayList<String> m = new ArrayList<>();
-			for (Component c : getComponents()) {
-				if (c.getClass() != HintTextArea.class)
-					continue;
-				HintTextArea h = (HintTextArea) c;
-				m.add(h.getText());
-			}
-			return m;
-		}
-	}
 }
